glasir_api/requirements.txt
fastapi
uvicorn[standard] # Includes standard dependencies like websockets, needed for development server
httpx
beautifulsoup4
lxml
pydantic
cachetools
aiofiles
orjson

sqlalchemy
databases[sqlite]
python-dotenv
alembic
fastapi-limiter
